# pkg/api/payment/service/v1/payment_service.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "payment/service/v1/payment_service.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "PaymentService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/payment/v1/create": {
      "post": {
        "summary": "Создание нового платежа",
        "description": "Этот метод позволяет создать новый платеж",
        "operationId": "PaymentService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/responsesCreateResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/requestsCreateRequest"
            }
          }
        ],
        "tags": [
          "payment"
        ]
      }
    },
    "/payment/v1/list": {
      "get": {
        "summary": "Получение списка платежей",
        "description": "Этот метод возвращает список всех платежей",
        "operationId": "PaymentService_List",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/responsesListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "description": "Количество записей на странице",
            "in": "query",
            "required": false,
            "type": "string",
            "format": "int64"
          },
          {
            "name": "cursor",
            "description": "Указатель на последний платеж",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "payment"
        ]
      }
    },
    "/payment/v1/{id}": {
      "get": {
        "summary": "Получение платежа по ID",
        "description": "Этот метод возвращает информацию о платеже по его уникальному идентификатору",
        "operationId": "PaymentService_Get",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/responsesGetResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "description": "id транзакции",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "payment"
        ]
      }
    }
  },
  "definitions": {
    "enumscryptocurrencyv1Currency": {
      "type": "string",
      "enum": [
        "CURRENCY_CRYPTO_UNSPECIFIED",
        "CURRENCY_BTC",
        "CURRENCY_TON",
        "CURRENCY_USDT"
      ],
      "default": "CURRENCY_CRYPTO_UNSPECIFIED",
      "title": "Список доступных криптовалют(BTC, TON, ...)"
    },
    "enumsfiatcurrencyv1Currency": {
      "type": "string",
      "enum": [
        "CURRENCY_FIAT_UNSPECIFIED",
        "CURRENCY_RUB"
      ],
      "default": "CURRENCY_FIAT_UNSPECIFIED",
      "title": "Список доступных фиатных валют(RUB)"
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "paymentstatusv1Status": {
      "type": "string",
      "enum": [
        "STATUS_UNSPECIFIED",
        "STATUS_CREATED",
        "STATUS_PENDING",
        "STATUS_SUCCESS",
        "STATUS_FAILED",
        "STATUS_UNKNOWN"
      ],
      "default": "STATUS_UNSPECIFIED",
      "title": "Статус оплаты(PENDING, SUCCESS, ...)"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "requestsCreateRequest": {
      "type": "object",
      "properties": {
        "paymentData": {
          "$ref": "#/definitions/v1Transaction"
        }
      },
      "title": "Создание транзакции перевода из фиатной валюты в криптовалюту"
    },
    "requestsGetRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id транзакции"
        }
      },
      "title": "Получить информацию о платежной транзакции"
    },
    "responsesCreateResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id транзакции"
        }
      }
    },
    "responsesGetResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "id транзакции"
        },
        "transaction": {
          "$ref": "#/definitions/v1Transaction",
          "title": "Встраиваем общую структуру для данных платежа"
        }
      },
      "title": "Получение данных о платежной транзакции по id"
    },
    "responsesListResponse": {
      "type": "object",
      "properties": {
        "payments": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/requestsGetRequest"
          },
          "title": "Список платежей"
        },
        "nextCursor": {
          "type": "string",
          "title": "Курсор для следующей страницы (если он есть)"
        }
      },
      "title": "Возвращает список платежей с пагинацией(cursor-based)"
    },
    "transactiontypev1Type": {
      "type": "string",
      "enum": [
        "TYPE_UNSPECIFIED",
        "TYPE_INVOICE",
        "TYPE_WITHDRAW"
      ],
      "default": "TYPE_UNSPECIFIED",
      "title": "Тип перевода(INVOICE, WITHDRAW)"
    },
    "transactionv1Operation": {
      "type": "object",
      "properties": {
        "account": {
          "type": "string",
          "title": "Реквизит: номер карты или адрес крипто-кошелька"
        },
        "money": {
          "$ref": "#/definitions/v1Money",
          "title": "Данные о сумме операции(amount, success_paid, fee)"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "Операция списания или зачисления"
    },
    "v1Money": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "title": "Введённая сумма"
        },
        "successPaid": {
          "type": "string",
          "title": "Сумма + комиссия"
        },
        "fee": {
          "type": "string",
          "title": "Комиссия (если есть)"
        },
        "crypto": {
          "$ref": "#/definitions/enumscryptocurrencyv1Currency",
          "description": "BTC, TON ..."
        },
        "fiat": {
          "$ref": "#/definitions/enumsfiatcurrencyv1Currency",
          "description": "RUB, ..."
        }
      },
      "title": "Универсальное представление денежек"
    },
    "v1Transaction": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "title": "payment_id - id транзакции"
        },
        "status": {
          "$ref": "#/definitions/paymentstatusv1Status",
          "title": "Текущий статус платежа"
        },
        "type": {
          "$ref": "#/definitions/transactiontypev1Type",
          "title": "Тип платежа"
        },
        "from": {
          "$ref": "#/definitions/transactionv1Operation",
          "title": "Операция списания (например, списание с yoomoney)"
        },
        "to": {
          "$ref": "#/definitions/transactionv1Operation",
          "title": "Операция зачисления (например, начисление на крипто-кошелек)"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "Обменная транзакция, включающая списание и зачисление"
    }
  }
}